Updated tasks 10, 11, 14, 23; added tasks 25 and 26.

+++ src/part4/task14/Multiplication.java
@@ -2,8 +2,8 @@ package part4.task14;
 
 public class Multiplication {
     public static void main(String[] args) {
-        int i = 1;
-        int multiplication = 1;
+        long i = 1L;
+        long multiplication = 1L;
         do {
             multiplication *= i;
             i++;
